refactor : 컴포넌트 명과 동일하게 파일명 변경
#228
